dired-rsync: fix shell quoting
PR #3 ends up double quoting arguments as we already do it as we
convert remote arguments from tramp format to be ready for the rsync.

We update rename the tramp helper and apply it to both src-files and
the destination. We also drop the unused :source and :dest parameters
from do-run-cmd. We can add them back later if we go on to use them.

Signed-off-by: Alex Bennée <alex@bennee.com>