dired-rsync: convert extract user to use tramp functions
This isn't a complete solution. The edge case of copying from remote
to remote relies on the first remote knowing the username to log in
for the second remote. Ideally we should resolve that locally with
whatever .ssh/config contains. For now just mark the limitation and
fallback to (getenv "USER") which might DTRT. Expand the test case
while we are at it.