dired-rsync: clean up process buffer
rsync is helpfully throwing in newlines for the benefit of the
terminal to show progress. There is no real need for us to keep all
that historical status data once we have scrapped what we want.

We can now remove the window-system hack which sets the 'mac coding
system as that was treating the symptom of not cleaning up lines.

Fixes: #36, #39